requirements.txt
pyyaml
pillow
numpy
pandas
matplotlib
opencv-python
scikit-image
scikit-learn
tqdm
nibabel